Refactored some collision logic to handle domination point trigger.

+++ capstone-game-cpp/Components/DominationPointController.cpp
@@ -5,7 +5,7 @@
 #include "Mathf.h"
 #include <Colors.h>
 
-void DominationPointController::Update(float milliseconds)
+void DominationPointController::Update(double milliseconds)
 {
 	if (isDominated)
 	{
